+++ front/package.json
@@ -10,7 +10,7 @@
     "@types/node": "^16.18.25",
     "@types/react": "^18.2.1",
     "@types/react-dom": "^18.2.2",
-    "axios": "^1.5.0",
+    "axios": "^1.5.1",
     "moment": "^2.29.4",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
@@ -20,7 +20,7 @@
     "socket.io-client": "^4.7.2",
     "typescript": "^4.9.5",
     "web-vitals": "^2.1.4",
-    "zustand": "^4.4.1"
+    "zustand": "^4.4.3"
   },
   "scripts": {
     "start": "react-scripts start",
@@ -47,6 +47,6 @@
     ]
   },
   "devDependencies": {
-    "@types/react-modal": "^3.16.0"
+    "@types/react-modal": "^3.16.1"
   }
 }
